requirements.txt
alembic==1.13.3
black==24.10.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.14
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2024.10.0
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-error-reporting==1.11.1
google-cloud-logging==3.11.3
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
idna==3.10
importlib_metadata==8.4.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
mccabe==0.7.0
mpmath==1.3.0
mypy==1.12.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==75.2.0
SQLAlchemy==2.0.35
sympy==1.13.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2
zipp==3.20.2
gunicorn==23.0.0